refractor UI refresher
