+++ src/features/farming/quarry/components/Iron.tsx
@@ -24,6 +24,7 @@ import { canMine } from "features/game/events/ironMine";
 import { miningAudio, miningFallAudio } from "lib/utils/sfx";
 import { HealthBar } from "components/ui/HealthBar";
 import { TimeLeftPanel } from "components/ui/TimeLeftPanel";
+import useUiRefresher from "lib/utils/hooks/useUiRefresher";
 
 const POPOVER_TIME_MS = 1000;
 const HITS = 3;
@@ -55,19 +56,7 @@ export const Iron: React.FC<Props> = ({ rockIndex }) => {
   const mined = !canMine(rock);
   const { setToast } = useContext(ToastContext);
 
-  const [_, setTimer] = React.useState<number>(0);
-  const setRecoveryTime = React.useCallback(() => {
-    setTimer((count) => count + 1);
-  }, []);
-
-  // refresh every second
-  useEffect(() => {
-    if (mined) {
-      setRecoveryTime();
-      const interval = window.setInterval(setRecoveryTime, 1000);
-      return () => window.clearInterval(interval);
-    }
-  }, [mined, setRecoveryTime]);
+  useUiRefresher({ active: mined });
 
   // Reset the shake count when clicking outside of the component
   useEffect(() => {
