+++ src/features/game/expansion/components/resources/Tree.tsx
@@ -33,6 +33,7 @@ import { LandExpansionTree } from "features/game/types/game";
 import { canChop } from "features/game/events/landExpansion/chop";
 import { Overlay } from "react-bootstrap";
 import { calculateBumpkinStamina } from "features/game/events/landExpansion/replenishStamina";
+import useUiRefresher from "lib/utils/hooks/useUiRefresher";
 
 const POPOVER_TIME_MS = 1000;
 const HITS = 3;
@@ -93,19 +94,7 @@ export const Tree: React.FC<Props> = ({ treeIndex, expansionIndex }) => {
 
   const chopped = !canChop(tree);
 
-  const [_, setTimer] = React.useState<number>(0);
-  const setRecoveryTime = React.useCallback(() => {
-    setTimer((count) => count + 1);
-  }, []);
-
-  // refresh every second
-  useEffect(() => {
-    if (chopped) {
-      setRecoveryTime();
-      const interval = window.setInterval(setRecoveryTime, 1000);
-      return () => window.clearInterval(interval);
-    }
-  }, [chopped, setRecoveryTime]);
+  useUiRefresher({ active: chopped });
 
   const displayPopover = async (element: JSX.Element) => {
     setPopover(element);
